requirements.txt
request==2.21.0
terminaltables==3.1.0
python-dotenv==0.10.3